Factory drawer

+++ src/sample/Factory/IPolygonDrawerFactory.java
@@ -4,5 +4,5 @@ import sample.View.IShapeDrawer;
 import sample.View.IView;
 
 public interface IPolygonDrawerFactory {
-    public IShapeDrawer createPolygonDrawer(Double[] points, IView view, Double[] rgb);
+    public IShapeDrawer createPolygonDrawer(Double[] points, IView view, int[] rgb);
 }
